% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_and_long.R
\name{aggregate_to_subclass_se}
\alias{aggregate_to_subclass_se}
\title{Aggregate molecule-level SE to subclass-level SE via rowData$subclass}
\usage{
aggregate_to_subclass_se(
  se_mol,
  fun = c("sum", "mean", "median", "proportion_within_class"),
  na_rm = TRUE
)
}
\arguments{
\item{se_mol}{Molecule-level SE with assay "abundance".}

\item{fun}{One of "sum","mean","median","proportion_within_class".}

\item{na_rm}{Remove NAs in aggregation.}
}
\value{
A SummarizedExperiment where rows are subclasses.
}
\description{
Aggregate molecule-level SE to subclass-level SE via rowData$subclass
}
