% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_and_long.R
\name{as_long_table}
\alias{as_long_table}
\title{Convert SE to tidy long table for plotting/statistics}
\usage{
as_long_table(
  se,
  level_label = "molecule",
  include_rowdata = c("name", "class", "subclass", "mz", "rt_min")
)
}
\arguments{
\item{se}{A SummarizedExperiment with assay "abundance".}

\item{level_label}{A label stored in output (e.g., "molecule" or "subclass").}

\item{include_rowdata}{RowData columns to keep.}
}
\value{
A tibble in long format.
}
\description{
Convert SE to tidy long table for plotting/statistics
}
