% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mztab_m.R
\name{mztab_se_to_lipidomics_se}
\alias{mztab_se_to_lipidomics_se}
\title{Convert mzTab-derived SE (from mztab_to_se()) into a lipidomics-like SE}
\usage{
mztab_se_to_lipidomics_se(
  x,
  sample_id_col = "sample_id",
  class_col = "class",
  name_priority = c("Metabolite name", "chemical_name", "SML_ID", "SMF_ID_REFS"),
  subclass_candidates = c("Ontology", "ontology", "subclass"),
  keep_original_sample_names = TRUE,
  make_unique_sample_ids = TRUE,
  use_rgoslin_if_no_ontology = TRUE,
  rgoslin_name_col = "Metabolite name",
  verbose = FALSE
)
}
\arguments{
\item{x}{Output list from \code{mztab_to_se()} OR a SummarizedExperiment (\code{x$se}).}

\item{sample_id_col}{Column in mzTab colData used as sample_id (default "sample_id").}

\item{class_col}{Column in mzTab colData used as class (default "class").}

\item{name_priority}{Candidate rowData columns used to create "Metabolite name".}

\item{subclass_candidates}{Candidate rowData columns used to create subclass (if any).}

\item{keep_original_sample_names}{If TRUE, keep original mzTab assay colnames in colData$mz_sample_col.}

\item{make_unique_sample_ids}{If TRUE, make sample_id unique.}

\item{use_rgoslin_if_no_ontology}{If TRUE, try rgoslin to infer subclass when no ontology-like info exists.}

\item{rgoslin_name_col}{Column used as input names for rgoslin (default "Metabolite name").}

\item{verbose}{If TRUE, prints messages about fallbacks.}
}
\value{
SummarizedExperiment
}
\description{
This produces an SE compatible with the structure of \code{load_lipidomics_se()}:
\itemize{
\item assays$abundance: numeric matrix
\item colData: DataFrame with sample_id and class
\item rowData: keeps mzTab annotations and ensures \verb{Metabolite name} and \code{subclass}
}
}
\details{
If no Ontology-like column exists in rowData, this function can try to infer
subclass from lipid names using the rgoslin R package.

NOTE: This version ACCEPTS PARTIAL rgoslin results:
\itemize{
\item parsed rows get inferred subclass
\item failed rows become "Unknown"
}
}
